Extract VotingCutoffField component to eliminate duplication
The original PR introduced a time-category Voting Cutoff block that
was a near-exact copy of the existing block used for yes/no and
preference polls (~70 lines of duplicated JSX). Extract a single
VotingCutoffField component and use it from both call sites. Unified
the custom date/time input IDs since the blocks are mutually exclusive.

Net change: -114 lines in page.tsx, +100 lines in the new component.
